backend/temporal-workers/requirements.txt
temporalio==1.9.0
httpx==0.28.1
python-dotenv==1.0.1
pydantic==2.10.6
gradio_client==1.10.2
aiohttp==3.12.13

